+++ source/main.cpp
@@ -20,7 +20,6 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#include <format>
 #include <fstream>
 #include <memory>
 #include "Machine/MachineFactory.h"
@@ -31,45 +30,30 @@ int main(void)
 {
 	try
 	{
+		// Open the configuration file, see the README for an explanation of each configuration option
 		std::ifstream fin(CONFIG_DIR"/config.json");
 		const nlohmann::json config = nlohmann::json::parse(fin);
-
-		// The machine to run Space Invaders on.
-		// Space Invaders runs at 60Hz with 2 interrupts per second, set the machine clock resolution accordingly.
-		// We require 4 interrupts, 2 for Space Invaders and 2 machine level interrupts for loading and saving.
-		// Ideally we would lock the interrupt service routine frequency to the clock resolution ("isrFreq":1),
-		// however, we need to spare some time for checking for load and save requests, so we bump the isrFreq down
-		// by ten percent ("isrFreq":0.9). One could lower it further, this would make it more responsive (0.9 should
-		// be good enough). Increasing it above 1 would make it slower and not respond to load/save requests.
-		// Set the ram/rom layout.
-		auto machine = MachEmu::MakeMachine(std::format(R"({{
-												"clockResolution":{},
-												"isrFreq":0.9,
-												"loadAsync":true,
-												"romOffset":0,
-												"romSize":8192,
-												"ramOffset":8192,
-												"ramSize":57343,
-												"runAsync":true,
-												"saveAsync":true}})",
-												(1000000000 / 60) / 2).c_str());
-		
-		//Create our custom Space Invaders memory controller.
+		// Create our custom Space Invaders machine
+		auto machine = MachEmu::MakeMachine(config["mach-emu"].dump().c_str());
+		// Create our custom Space Invaders memory controller.
 		auto memoryController = std::make_shared<SpaceInvaders::MemoryController>();
-		//Create our custom Space Invaders I/O controller based on a specific configuration.
-		auto ioController = std::make_shared<SpaceInvaders::SdlIoController>(memoryController, config);
+		// Create our custom Space Invaders I/O controller based on a specific configuration.
+		auto ioController = std::make_shared<SpaceInvaders::SdlIoController>(memoryController, config["space-invaders"]["io"]);
+		
 		// Load the ROM into memory with the following layout
 		memoryController->Load(ROMS_DIR"/invaders-h.bin", 0x0000); // invaders-h 0000-07FF
 		memoryController->Load(ROMS_DIR"/invaders-g.bin", 0x0800); // invaders-g 0800-0FFF
 		memoryController->Load(ROMS_DIR"/invaders-f.bin", 0x1000); // invaders-f 1000-17FF
 		memoryController->Load(ROMS_DIR"/invaders-e.bin", 0x1800); // invaders-e 1800-1FFF
+		// Load the memory layout into the machine
+		machine->SetOptions(config["space-invaders"]["memory"].dump().c_str());
 		// Load our controllers into the machine.
 		machine->SetMemoryController(memoryController);
 		machine->SetIoController(ioController);
 		// Will be called from a different thread
-		// Need to make a copy if you want to hold the json string longer than the scope of this function
 		machine->OnSave([](const char* json)
 		{
+			// Need to make a copy of the json if you want to hold the json string longer than the scope of this function
 			std::ofstream fout(ROMS_DIR"/spaceInvaders.json", std::ios::trunc);
 			fout.exceptions(fout.failbit);
 			fout.write(json, strlen(json));
@@ -88,8 +72,11 @@ int main(void)
 			return loadJson.c_str();
 		});
 
+		// Run the machine asynchronously, the machine now owns the controllers and they should not be accessed
 		machine->Run(0x00);
+		// Run the io event loop until the 'q' key is be pressed or the window is closed
 		ioController->EventLoop();
+		// Wait for the machine to finish, once complete the controllers can be accessed safely
 		machine->WaitForCompletion();
 	}
 	catch (const std::exception& e)