+++ CHANGELOG.md
@@ -1,10 +1,13 @@
-0.4.0
+0.4.0 [1/05/24]
 * Updated to mach-emu-1.5.1.
 * Added `IMachine::Load` support.
 * Added `IMachine::Save` support.
+* Added support for Space Invaders Part II Deluxe.
+* Moved all required options to the config file.
+* Updated the README to document the config options.
 
 0.3.1 [15/04/24]
-* Fixed a deadlock on quit when the interrupt service routine.
+* Fixed a deadlock on quit when the interrupt service routine
   frequency is greater than 0.
 
 0.3.0 [21/03/24]